Extract InspectionItemRow component
